requirements.txt
langsmith==0.3.50
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
pymongo==4.13.2
fastapi==0.116.1
pydantic==2.11.7
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-dotenv==1.1.1